Added code for Login Page Session Values

+++ src/main/java/com/ecrops/repo/DistrictRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.ecrops.entity.Districts;
 
-public interface DistrictRepo  extends JpaRepository<Districts, Long>{
+public interface DistrictRepo  extends JpaRepository<Districts, Integer>{
 	
 	@Query(value="select wbdcode ,wbmcode,type_user from user_registration   where  userid=:userId",nativeQuery=true)
 	String getDistByDcodeAndUserId(String userId);
